requirements.txt
beir==2.1.0
faiss_cpu==1.10.0
langchain==0.3.21
langchain_chroma==0.2.2
langchain_community==0.3.20
langchain_core==0.3.49
langchain_openai==0.3.11
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
PyPDF2==3.0.1
python_docx==1.1.2
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence_transformers==3.4.1
streamlit==1.39.0
tiktoken==0.8.0
torch==2.6.0
tqdm==4.66.5
transformers==4.49.0
